requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beanie==1.25.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.7.0
Brotli==1.1.0
bunnet==1.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
cryptography==42.0.4
cssselect==1.2.0
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
execnet==2.0.2
fastapi==0.109.2
fasteners==0.19
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kaitaistruct==0.10
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.3.2
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.9.13
outcome==1.3.0.post0
packaging==23.2
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
pdbp==1.5.0
pillow==10.2.0
platformdirs==4.2.0
playwright==1.41.2
pluggy==1.4.0
psutil==5.9.8
py==1.11.0
pyasn1==0.5.1
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
pydivert==2.1.0
pyee==11.0.1
PyGetWindow==0.0.9
Pygments==2.17.2
pymongo==4.6.1
PyMsgBox==1.0.9
pynose==1.4.8
pyOpenSSL==24.0.0
pyotp==2.9.0
pyparsing==3.1.1
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.0.0
pytest-html==2.0.1
pytest-metadata==3.1.0
pytest-ordering==0.6
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-dotenv==1.0.1
python-multipart==0.0.7
pytweening==1.2.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
sbvirtualdisplay==1.3.0
selectolax==0.3.20
selenium==4.17.2
selenium-wire==5.1.0
seleniumbase==4.23.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.36.3
tabcompleter==1.3.0
toml==0.10.2
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
ujson==5.9.0
undetected-chromedriver==3.5.5
urllib3==2.2.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
zstandard==0.22.0
